* 'master' of https://github.com/PRAISE-group/smt-with-llms:
  added brackets in z3_to_c in z3utils.py
  minor
  Dynamic variable type detection in fuzzing file creation

+++ benchmarks/BV-benchamrks-unsat/bvisxdigit-32/isxdigit.c
@@ -1,6 +1,6 @@
 #include "isxdigit.h"
 
-uint8_t isxdigit__cb(uint8_t c)
+uint32_t isxdigit__cb(uint32_t c)
 {
   if (c >= '0') {
     if (c <= '9') {
@@ -19,7 +19,7 @@ uint8_t isxdigit__cb(uint8_t c)
   return 0;
 }
 
-uint8_t foo_cb(uint8_t c)
+uint32_t foo_cb(uint32_t c)
 {
   return c & 1u;
 }
